requirements.txt
toml
RPi.GPIO
cachetools
mutagen
pillow
